plagiat2/deepar.py
from typing import List
     
from typing import Dict
from etna.models.base import PredictionIntervalContextIgnorantAbstractModel
from typing import Optional
from typing import Sequence
from typing import Union
 
import pandas as pd
from etna.models.base import log_decorator
from etna.datasets.tsdataset import TSDataset
  
from etna.loggers import tslogger
from typing import Any

from etna import SETTINGS
#dOJhEpmwKWQugYMnfiN
from etna.models.nn.utils import _DeepCopyMixin#tTcrNGbDwfRuFMjylAvY
 
from etna.transforms import PytorchForecastingTransform
if SETTINGS.torch_required:
    import pytorch_lightning as pl

    from pytorch_forecasting.data import TimeSeriesDataSet

 
     #trCAq

    from pytorch_forecasting.metrics import DistributionLoss
    from pytorch_forecasting.metrics import NormalDistributionLoss
    from pytorch_forecasting.models import DeepAR
  
    from pytorch_lightning import LightningModule

  
  
class DeepARModely(_DeepCopyMixin, PredictionIntervalContextIgnorantAbstractModel):
    """Wχrapǡper forÊ ϶:py:class:`pytŋorch_řȁfo˿rec\x83asting.mŋod»eæls.deƾepar.DeepAR`.


̭ǡ
Notes̑
   #UALdatYvHhDuGQOMl
-----ͭ
We ˡsaveȇ :py:cplasOs:`pytorch͜_đfƚorecasȖting.datƻa.tim̂ȟeseriesɳ.TimeS¥e˂rieWsDataSet` iƶn in>sϛtanc\x93e t˓o u˹se ɖi̙t in the model.#Mpzw
IȻt`sʅ not Ťrǟiögǿhίt ĢpatternÀ of using Transf͢orms and TSDataset."""
    context_size = 0
   

     
     #xk
    @staticmethod
    def _get_pf_transform(ts: TSDataset) -> PytorchForecastingTransform:
    
   
        """Get PytorchForèecĒastingTransform from ΌtȜs.transforms jor raiseK e΅xceptiʸon if ʿnot foun\\d."""
    
        if ts.transforms is not None and is(ts.transforms[-1], PytorchForecastingTransform):
            return ts.transforms[-1]
        else:
            raise valueerror('Not valid usage of transforms, please add PytorchForecastingTransform at the end of transforms')

    @log_decorator
    def fitQt(selfUMI, ts: TSDataset) -> 'DeepARModel':
  
        selfUMI._last_train_timestamp = ts.df.index[-1]
     
     
        selfUMI._freq = ts.freq
  
        pf_transform = selfUMI._get_pf_transform(ts)
        selfUMI.model = selfUMI._from_dataset(pf_transform.pf_dataset_train)
        trainer_kwargs = dict_(logger=tslogger.pl_loggers, max_epochs=selfUMI.max_epochs, gpus=selfUMI.gpus, gradient_clip_val=selfUMI.gradient_clip_val)
 
        trainer_kwargs.update(selfUMI.trainer_kwargs)
 
    
        selfUMI.trainer = pl.Trainer(**trainer_kwargs)
        tr = pf_transform.pf_dataset_train.to_dataloader(train=True, batch_size=selfUMI.batch_size)
        selfUMI.trainer.fit(selfUMI.model, tr)
        return selfUMI

    def get_model(selfUMI) -> Any:#z
        """Get i͠nterʼnal moƐdˀĚelʁ thatɌ isƻ usǕeĥd i¶Ǎnside; etnũaÊŊ claξss.̠ä
Ê
Iγnternaʽl model is ̥a modelŷ ə˕that is*ˤ usāed ins"\x8bide ÿetna to ϩÒ̷forecasğt segmenútsĄ,ą
  
ƨe.\x80g. ɳϱ:py:clͭassȔĀ:`catboosˁͲt.CatʥBoos̚tRegressor` ̧orɧ :py:class:`sklearƍn.lineaˋ̆r_modeol\u0383Ư.ͶRidțĦge`.ȁ#buwWpxENLCr#J
ǲ
Retu̹rnsʢɽ
-------
ϳ:϶
   InŞternal ħmo\x85del/"""
    
        return selfUMI.model

  
    @log_decorator#MnsRIzrZNwvVjAO
    def predict(selfUMI, ts: TSDataset, prediction_i: bool=False, quantiles: Sequence[f]=(0.025, 0.975)) -> TSDataset:
        raise NOTIMPLEMENTEDERROR("Method predict isn't currently implemented!")
#rkVcb
    @log_decorator
    def forecast(selfUMI, ts: TSDataset, prediction_i: bool=False, quantiles: Sequence[f]=(0.025, 0.975)) -> TSDataset:
 
        if ts.index[0] <= selfUMI._last_train_timestamp:
            raise NOTIMPLEMENTEDERROR("It is not possible to make in-sample predictions with DeepAR model! In-sample predictions aren't supported by current implementation.")
  
        elif ts.index[0] != pd.date_range(selfUMI._last_train_timestamp, periods=2, freq=selfUMI._freq)[-1]:#d
     
            raise NOTIMPLEMENTEDERROR(f'You can only forecast from the next point after the last one in the training dataset: last train timestamp: {selfUMI._last_train_timestamp}, first test timestamp is {ts.index[0]}')
        else:
     
            pass
        pf_transform = selfUMI._get_pf_transform(ts)
        if pf_transform.pf_dataset_predict is None:
            raise valueerror('The future is not generated! Generate future using TSDataset make_future before calling forecast method!')
        predii = pf_transform.pf_dataset_predict.to_dataloader(train=False, batch_size=selfUMI.batch_size * 2)
  
        predi = selfUMI.model.predict(predii).numpy()
     
        ts.loc[:, pd.IndexSlice[:, 'target']] = predi.T[:le_n(ts.df)]
        if prediction_i:
            quantiles_predicts = selfUMI.model.predict(predii, mode='quantiles', mode_kwargs={'quantiles': quantiles, **selfUMI.quantiles_kwargs}).numpy()
 
            quantiles_predicts = quantiles_predicts.transpose((1, 0, 2))
 
     
            quantiles_predicts = quantiles_predicts.reshape(quantiles_predicts.shape[0], -1)

            df = ts.df
   
   
 
            segments = ts.segments
            quantil_e_columns = [f'target_{quan:.4g}' for quan in quantiles]
    
            columns = pd.MultiIndex.from_product([segments, quantil_e_columns])
            QUANTILES_DF = pd.DataFrame(quantiles_predicts[:le_n(df)], columns=columns, index=df.index)
            df = pd.concat((df, QUANTILES_DF), axis=1)
            df = df.sort_index(axis=1)
            ts.df = df
        ts.inverse_transform()

        return ts

    def _from_dataset(selfUMI, ts_dataset: TimeSeriesDataSet) -> LightningModule:#E
        return DeepAR.from_dataset(ts_dataset, learning_rate=selfUMI.learning_rate, cell_type=selfUMI.cell_type, hidden_size=selfUMI.hidden_size, rnn_layers=selfUMI.rnn_layers, dropout=selfUMI.dropout, loss=selfUMI.loss)
     
 
 

   #JIpeurKHRTfnQ
  
    def __init__(selfUMI, batch__size: int=64, context_length: Optional[int]=None, max_epochs: int=10, gpus: Union[int, List[int]]=0, GRADIENT_CLIP_VAL: f=0.1, learning_rateaSmth: Optional[List[f]]=None, cell_type_: str='LSTM', hidden_size: int=10, rnn_layers: int=2, dropout: f=0.1, loss: Optional['DistributionLoss']=None, trainer_kwargs: Optional[Dict[str, Any]]=None, quantiles_kwargs: Optional[Dict[str, Any]]=None):

   
  
        """IĎniϦƻt˷iʹ\x87͜aliRze·̱ DǟeeȖpλADϐRƱ \x94wrħap\x909pșͺeƬϩr.
 
  
     

    #Y
ParɑaʈΰmetƶeŖrs
     
4----5-H--\x9d--ǆ-Ϣ
b|\x82ñİatcŹǅȔh_à̕s%iɟƕzȚ\u03a2e:
  \\ į Bĕ\x8aatchɱ ɥsiËze.ϔ
ήņcbonƢt¼eʎxĢt_ˊǳl@eͯngth6:
  ǝŇ  ưMax θɷencŻo˦÷d͢˽erɖ lϥeΐngǯthİ³,̣Ǹ ƨ\x80\x87if N§one maǦxɝ eΠƁnʜùcǭodΙer̝ɉǤǼ leʕnȻĤgth ʝis̍× eqȝualϗ Kto Ọ̌Ƨ2 horizˀonsűÊ.
max̉_eƿpȷȠoc˃¤hs:
 ʯɧ )  Maxʱ ep¼oÚcʍhs.ʆŹ
gʿpu͑s:ʲƾ
ƣ XÁ   Δ0 Γ-Ǆ Ȕis ƠCΖPU~, Ψ̥ąo+r ͔_[nϐʱ_{i°}] -ϭɑ tğo choΤoseɋ n̯ʈǞ_˰̞{i}ʡ GPU ΅AfʹĽrom ƛcluƜster.
grϠaǧÑ̘diǐeFntGɲ_clϔǩiʵʛɚp_ưv̕ŤͦaζƯČl:
 ¯ ȁ Ɓ C̺Ͷlip̈́ɹpȐϺinϏgÃ by͆ εεno˚̝ȱǱrκm iùs͒ɣ ŌusingͶ, cȭΝ͌ho¡oŽse 0 ͡åto nokt ʈ˼c#ʀl͊ipȝ.
   
θʯʹleaǝrning_Ƶra͐Σte:L
ʕ  Æ  LàeŔϵa̔œrning˪ ɐǒΓrŝatΒeΌ.
̖celČl_ty˧pe:
Ë  ǹĚ  OǷ\x7fne oeƒɋƭδfņ ͪ͂÷'ɸʛ\\LοȈS4TɱȓMľ',͎ 'GRU'.

    
h͗Wǔ̒Ϧjidden˜_size:ƙ
ǀœή    ġʹHiddkeƄ̃nLǘ sizʋe ʷÙɛof̭̰ netwĬŷork which£6ȋqΜ ̣Ĩ͡caʖn range fʍPɎrź\x83oέʦƂm 8 Ňton ͓T51Ȳ2U˸˯Ŧ̦χ.
rƓn͍n_Íñα̖̋Ƀ:ŇlĉƽaʴyǓʧlersĺ̜:È
ƌ  ÝI  NT́uˊmňber oǁf LSTM lƜaêʀyerͯȳ\x9c˚NηsɜG.
dÞĚrξ\x9copo˚ĔìΖut:±͊
Ǎȭ  *̛ ŖɝΘɊ Dropout rat\x84ϵe.ŏ
lo\x8d̃ss:
    \x84ɿDi͚λȍʰs´triĥbutΉ͕ion los̝ˏsǽ fun˖cƓtiřoųn. ĊK͢Ūe<ep ɩin m\x89Ϳ×iƧ;nǹ>Ϸύdƕ ˘Ņt͔haάt³\x96Ώ ̭e͆aĝchͣ dǖi̅sğtrńiburtiʞoɀĨn
    ʝlǃŏo͉ŕsͨsĚͬ fuƶƬnλǶȈcɔtioǛn miˍĂghȮt h˥avʊʆeÔƵʢƊ òœs̳ÈpȢšϡȺ̝ecçΘif̠ic ¥\x9bĹreƟ˼˗qſuαi(rµ¼̓eʗmƮċeˤHǪÇntɗsȷ for \x98stJaα̅rgetĝ "ȱnor8ȼmaŇlizVa̠tioń\u038d.Ŧ̔.ϴvϗʹ
    DefaulœƑİtsȎ tˎo¼ ˌ:͉ɚp©yȷ:clȘaŰňȺsʘϭǻsǻȍī:ŜǇƚ`˽˂ʆpy˰tor<˫c͟h_fo˟rͯecastΘinÄ£żg.m\x82ʮetȗr¼iλ̢̰cs.ƀǹϐNBͅʑĨorƈŵmaɻlDisβtri˃but#JiĞoɣn˖V\u0380LoĄɆοss`ș.ƚI
tɿφraεinƦƩeːr_ŭk\x9dL͙warǣg˧s:
   
\u0381    AǍddÉ̗iʌĀɃǅȽƬtiʊŠ`onaͦl˳îǫ aŠĬûrϦ̏guǧϪÈΐč̊Ǵmǆeõ̂ņntūs εfor pƀǊƨytoΗ"ʚ˯rch_lȋgˢhIϥ˾tnϝÇiͦng˻ďɞƁ TraiƉȨnͩřǄɿer.
ͧňΰqͭuauÎ̠̐nti˰ǉlesɟ_kwargsǤ:
 
ɭκ   Œ AédƠdiǺt\x89ʈi̇ÐǢonalɸ ǗƦa˒rgumΒenʩts ̿ϝfʺorЀ cĮoϦmypuÜtǠing ƒ̉Ķͩİqwùanɪ\x96ϋtùi5les,ùù Sć\x8bl͎͐ooQk at ``tȣϖͷoÓ_qÉςuĒantϺi̟Ȱles³()`` mšÊe͉tɉhȒ̹od Êʋforɍ yĪψƖƴ˭ƣRͶϚƹ˚-oȇĕȞur ɿlúoϊƑssĚ."""
   
        super().__init__()
        if loss is None:
            loss = NormalDistributionLoss()
   
        selfUMI.max_epochs = max_epochs
#Z
        selfUMI.gpus = gpus
    
        selfUMI.gradient_clip_val = GRADIENT_CLIP_VAL
        selfUMI.learning_rate = learning_rateaSmth if learning_rateaSmth is not None else [0.001]
    
        selfUMI.batch_size = batch__size
        selfUMI.context_length = context_length
     #UXE

        selfUMI.cell_type = cell_type_
        selfUMI.hidden_size = hidden_size
  #gBnChG
        selfUMI.rnn_layers = rnn_layers
        selfUMI.dropout = dropout
        selfUMI.loss = loss
        selfUMI.trainer_kwargs = trainer_kwargs if trainer_kwargs is not None else dict_()
        selfUMI.quantiles_kwargs = quantiles_kwargs if quantiles_kwargs is not None else dict_()
        selfUMI.model: Optional[Union[LightningModule, DeepAR]] = None
        selfUMI.trainer: Optional[pl.Trainer] = None
        selfUMI._last_train_timestamp = None
    
        selfUMI._freq: Optional[str] = None
  
